+++ Sora/Sora.cs
@@ -72,6 +72,7 @@ public class Sora : IDisposable
     {
         OPUS,
     }
+
     // SimulcastRid のためのパラメータ
     public enum SimulcastRidType
     {
@@ -87,6 +88,17 @@ public class Sora : IDisposable
         R0,
         R1,
         R2,
+    }
+    public enum DegradationPreference
+    {
+        // degradation preference を無効にします
+        Disabled,
+        // フレームレートを優先して維持し、帯域不足時は主に解像度を下げます
+        MaintainFramerate,
+        // 解像度を優先して維持し、帯域不足時は主にフレームレートを下げます
+        MaintainResolution,
+        // 解像度とフレームレートの両方をバランスよく下げます。
+        Balanced,
     }
     public class DataChannel
     {
@@ -181,7 +193,7 @@ public class Sora : IDisposable
         Internal,
         CiscoOpenH264,
         IntelVpl,
-        NvidiaVideoCodecSdk,
+        NvidiaVideoCodec,
         AmdAmf,
     }
     static string VideoCodecImplementationToString(VideoCodecImplementation implementation)
@@ -194,8 +206,8 @@ public class Sora : IDisposable
                 return "cisco_openh264";
             case VideoCodecImplementation.IntelVpl:
                 return "intel_vpl";
-            case VideoCodecImplementation.NvidiaVideoCodecSdk:
-                return "nvidia_video_codec_sdk";
+            case VideoCodecImplementation.NvidiaVideoCodec:
+                return "nvidia_video_codec";
             case VideoCodecImplementation.AmdAmf:
                 return "amd_amf";
             default:
@@ -212,8 +224,8 @@ public class Sora : IDisposable
                 return VideoCodecImplementation.CiscoOpenH264;
             case "intel_vpl":
                 return VideoCodecImplementation.IntelVpl;
-            case "nvidia_video_codec_sdk":
-                return VideoCodecImplementation.NvidiaVideoCodecSdk;
+            case "nvidia_video_codec":
+                return VideoCodecImplementation.NvidiaVideoCodec;
             case "amd_amf":
                 return VideoCodecImplementation.AmdAmf;
             default:
@@ -311,13 +323,13 @@ public class Sora : IDisposable
         }
 
         // 可能な限り HWA を利用する VideoCodecPreference を返す
-        // 優先度的には Intel VPL > Nvidia Video Codec SDK > Internal となる
+        // 優先度的には Intel VPL > AMD AMF > Nvidia Video Codec > Internal となる
         public static VideoCodecPreference GetHardwareAcceleratorPreference(VideoCodecCapability capability)
         {
             // Merge は同じコーデックを上書きするので、優先度が高いのを後でマージすることで正しい優先度になる
             var preference = new VideoCodecPreference();
             preference.Merge(CreateFromImplementation(capability, VideoCodecImplementation.Internal));
-            preference.Merge(CreateFromImplementation(capability, VideoCodecImplementation.NvidiaVideoCodecSdk));
+            preference.Merge(CreateFromImplementation(capability, VideoCodecImplementation.NvidiaVideoCodec));
             preference.Merge(CreateFromImplementation(capability, VideoCodecImplementation.AmdAmf));
             preference.Merge(CreateFromImplementation(capability, VideoCodecImplementation.IntelVpl));
             return preference;
@@ -372,6 +384,7 @@ public class Sora : IDisposable
         public string VideoAv1Params = "";
         public string VideoH264Params = "";
         public int VideoBitRate = 0;
+        public DegradationPreference? DegradationPreference;
         // デバイスから録音する代わりに Sora.ProcessAudio() で指定したデータを録音データとして利用するかどうか
         public bool UnityAudioInput = false;
         // 再生データをデバイスで再生する代わりに Sora.OnHandleAudio コールバックで再生データを受け取るようにするかどうか
@@ -602,6 +615,24 @@ public class Sora : IDisposable
         cc.video_av1_params = config.VideoAv1Params;
         cc.video_h264_params = config.VideoH264Params;
         cc.video_bit_rate = config.VideoBitRate;
+        if (config.DegradationPreference.HasValue)
+        {
+            switch (config.DegradationPreference.Value)
+            {
+                case DegradationPreference.Disabled:
+                    cc.SetDegradationPreference(SoraConf.Internal.DegradationPreference.DISABLED);
+                    break;
+                case DegradationPreference.MaintainFramerate:
+                    cc.SetDegradationPreference(SoraConf.Internal.DegradationPreference.MAINTAIN_FRAMERATE);
+                    break;
+                case DegradationPreference.MaintainResolution:
+                    cc.SetDegradationPreference(SoraConf.Internal.DegradationPreference.MAINTAIN_RESOLUTION);
+                    break;
+                case DegradationPreference.Balanced:
+                    cc.SetDegradationPreference(SoraConf.Internal.DegradationPreference.BALANCED);
+                    break;
+            }
+        }
         cc.unity_audio_input = config.UnityAudioInput;
         cc.unity_audio_output = config.UnityAudioOutput;
         cc.audio_recording_device = config.AudioRecordingDevice;